+++ app/code/community/Ffuenf/DevTools/Block/Adminhtml/Widget/Grid/Backup.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * Ffuenf_DevTools extension.
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the MIT License
+ * that is bundled with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/mit-license.php
+ *
+ * @category   Ffuenf
+ *
+ * @author     Achim Rosenhagen <a.rosenhagen@ffuenf.de>
+ * @copyright  Copyright (c) 2016 ffuenf (http://www.ffuenf.de)
+ * @license    http://opensource.org/licenses/mit-license.php MIT License
+ */
+
+class Ffuenf_DevTools_Block_Adminhtml_Widget_Grid_Backup extends Mage_Adminhtml_Block_Widget_Grid_Container
+{
+    public function __construct()
+    {
+        $this->_controller = 'backup';
+        $this->_blockGroup = 'devtools';
+        $this->_headerText = Mage::helper('catalog')->__('Backup');
+        parent::__construct();
+        $this->_removeButton('add');
+    }
+}