+++ app/code/community/Ffuenf/DevTools/Block/Sales/Order/Info.php
@@ -1,0 +1,41 @@
+<?php
+/**
+ * Ffuenf_DevTools extension.
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the MIT License
+ * that is bundled with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/mit-license.php
+ *
+ * @category   Ffuenf
+ *
+ * @author     Achim Rosenhagen <a.rosenhagen@ffuenf.de>
+ * @copyright  Copyright (c) 2016 ffuenf (http://www.ffuenf.de)
+ * @license    http://opensource.org/licenses/mit-license.php MIT License
+ */
+
+class Ffuenf_DevTools_Block_Sales_Order_Info extends Mage_Sales_Block_Order_Info
+{
+
+    /**
+     * Set the original module name avoid breaking translations
+     */
+    public function __construct()
+    {
+        parent::__construct();
+        $this->setModuleName('Mage_Sales');
+    }
+
+    /**
+     * Remove Link
+     */
+    public function removeLink($name)
+    {
+        if (isset($this->_links[$name])) {
+            unset($this->_links[$name]);
+        }
+        return $this;
+    }
+}