+++ CHANGELOG.md
@@ -2,6 +2,38 @@
 
 This file is used to list changes made in each version of Ffuenf_DevTools.
 
+## 2.1.9 (March 12, 2016)
+
+* update travis / MageTestStand
+* [docs] minor updates
+* fix helper
+* re-organize menu / navigation
+* fix date input
+
+## 2.1.8 (February 25, 2016)
+
+* update travis build-matrix
+* [docs] copyright notice
+
+## 2.1.7 (December 20, 2015)
+
+* fix date input in backend configuration
+* add date type config for resave-products shell script (created_at / updated_at)
+* fix dependency to [Ffuenf_Common](https://github.com/ffuenf/Ffuenf_Common)
+* code-style
+* update travis build-matrix
+
+## 2.1.6 (November 22, 2015)
+
+* add uninstallation via [Ffuenf_MageTrashApp](https://github.com/ffuenf/Ffuenf_MageTrashApp)
+
+## 2.1.5 (November 21, 2015)
+
+* add dependency to [Ffuenf_Common](https://github.com/ffuenf/Ffuenf_Common)
+* common extension structure
+* add more tests
+* add note about inherited extensions
+
 ## 2.1.4 (November 8, 2015)
 
 * switch back to official builds for versions > 1.9.0.0 in travis tests